requirements.txt
Django==3.0.8
django-dotenv==1.4.2
djangorestframework==3.12.2
djangorestframework-api-key==2.0.0
drf-nested-routers==0.92.5
psycopg2==2.8.6
psycopg2-binary==2.8.6